Added tests Pulled Ui out of GameManager excluded GameViewManager from Sonar Small change to GameManager
